changed balloon background and name background
The old balloon style was too rigid; the gap in the balloon's border looked nice but makes it impossible to adjust npc name positioning without having to update the background image as well. This one doesn't look as nice, but it's much easier to work with. Adjusted npc name background to compensate so the name text is still easy to read.